Track generated sources in tree
This removes use of the submodule with sources and instead commits the
generated sources directly into the tree (we will add a GH action
keeping them up-to-date in a follow-up).

This commit adds files generated by

```console
$ tree-sitter generate
```

We add all currently generated bindings and other generated files. This
e.g., adds bindings for Python (for easier use in zeek/zeekscript) and
also Rust.