Remove testing badge from README
The README is displayed e.g., on pypi.org and contained a badge to the
results of a run against the full Zeek source tree. If we display it
that prominently we should make sure to keep test passing.

Currently the parser has a number of issues parsing general Zeek script
code, especially around preprocessor constructs (e.g., nested preproc
statements, pragmas in general). Drop the status build badge, at least
until this is fixed.