Remove explicit installation of tree-sitter CLI in CI check job
The GH action `tree-sitter/setup-action/cli` defaults to using the
latest version of tree-sitter while we already already capture the
tree-sitter version we build against in `package.json`. This duplication
was not only unnecessary, but it also lead to version conflicts when
checking whether generated files were up-to-date. This instead uses the
tree-sitter CLI version as encoded for npm.